restructure integration tests
Divide integrations tests into folders. Divide test helpers into folders. Delete some useless tests.
