+++ src/Frontend/Components/ManualPackagePanel/__tests__/ManualPackagePanel.test.tsx
@@ -7,7 +7,7 @@ import { fireEvent, screen } from '@testing-library/react';
 import React from 'react';
 import { renderComponentWithStore } from '../../../test-helpers/render-component-with-store';
 import { ManualPackagePanel } from '../ManualPackagePanel';
-import { getParsedInputFileEnrichedWithTestData } from '../../../test-helpers/test-helpers';
+import { getParsedInputFileEnrichedWithTestData } from '../../../test-helpers/general-test-helpers';
 import { setSelectedResourceId } from '../../../state/actions/resource-actions/audit-view-simple-actions';
 import { loadFromFile } from '../../../state/actions/resource-actions/load-actions';
